Squashed commit of the following:
commit 39d027c2c86572820cea291119bb368d2a1e5ea3
Author: SGrottel <webmaster@sgrottel.de>
Date:   Mon Jul 4 21:01:47 2022 +0200

    Option to isolate OpenHereSettings app in a subfolder

commit ed22285c8cd3a13dcc816cd453f0e899a471a929
Author: SGrottel <webmaster@sgrottel.de>
Date:   Mon Jul 4 20:38:13 2022 +0200

    ToolboxApp can open SettingsApp

commit 34dd39e02bf8a6bbf05b2a224a62126ed17e69fb
Author: SGrottel <webmaster@sgrottel.de>
Date:   Mon Jul 4 20:30:25 2022 +0200

    Typo in README.md

commit dc027a33ef00b9f3e7ed9e86eef85732c85a2f65
Merge: 76358be 83422f8
Author: SGrottel <webmaster@sgrottel.de>
Date:   Mon Jul 4 20:28:12 2022 +0200

    Merge branch 'develop' into settings-ui-app

commit 76358be198f01f66db84125eb29ec2fc982280b0
Author: SGrottel <webmaster@sgrottel.de>
Date:   Mon Jul 4 20:27:34 2022 +0200

    SettingsApp SelectIconDialog implemented

commit 4f1da1ea304abb6e49347df2cfc0d0b56c655abc
Author: SGrottel <webmaster@sgrottel.de>
Date:   Mon Jul 4 13:14:10 2022 +0200

    Class IconLibrary now explicitly available; preparation for icon selection UI

commit 34d275821054a0cbcf9bd120c1ccd24c3e4758de
Author: SGrottel <webmaster@sgrottel.de>
Date:   Sun Jul 3 17:42:34 2022 +0200

    ToolInfoView.CombinedIcon parsing (not perfect, but ok for now)

commit 2d32f1d3f30d8e7eb7e2ff2a2a422d7e8847cef7
Author: SGrottel <webmaster@sgrottel.de>
Date:   Sun Jul 3 17:34:12 2022 +0200

    Started icon selection UI (works for ico files only, yet)

commit ff4278ccee8518fc1e188fd8f34243e3c1c1ed75
Author: SGrottel <webmaster@sgrottel.de>
Date:   Sun Jul 3 17:14:59 2022 +0200

    Implemented "Save" (to user settings)

commit 754233450e3809af5646be2e8c325df89b087c0e
Author: SGrottel <webmaster@sgrottel.de>
Date:   Sun Jul 3 17:03:39 2022 +0200

    ImportExport.FormatVersion check added

commit 8180a2e9b9b1b13c91993d20ec9f1c1924abe925
Author: SGrottel <webmaster@sgrottel.de>
Date:   Sun Jul 3 16:58:33 2022 +0200

    Implemented full data import&export

commit 3e762e124200b017c6afe4dd531f639995b0d987
Author: SGrottel <webmaster@sgrottel.de>
Date:   Sun Jul 3 16:31:31 2022 +0200

    Icon resource to indicate icon load failure

commit 99cc885126edeffb437aa2650779b73a62788f53
Author: SGrottel <webmaster@sgrottel.de>
Date:   Sat Jul 2 20:28:25 2022 +0200

    Browse buttons for executable and working directory

commit 9a9f677e17c068707b5c050400a19e85e845b280
Author: SGrottel <webmaster@sgrottel.de>
Date:   Sat Jul 2 19:41:20 2022 +0200

    Most UI elements for ToolInfo:StartConfig editing

commit 29efe8c050d1676df192564f9399282ab62431d7
Author: SGrottel <webmaster@sgrottel.de>
Date:   Sat Jul 2 17:26:14 2022 +0200

    Drag&Drop reordering in ToolsList ListView implemented

commit 1eed0869b8e2a199116dc79e27336ba995d164a6
Author: SGrottel <webmaster@sgrottel.de>
Date:   Sat Jul 2 16:39:31 2022 +0200

    Newlines newlines newlines newlines

commit 8e34c7941e7b00fde393ffa1fa5edb47d558008f
Merge: 9078d30 22ca8b1
Author: SGrottel <webmaster@sgrottel.de>
Date:   Sat Jul 2 16:35:45 2022 +0200

    Merge branch 'develop' into settings-ui-app

    # Conflicts:
    #	ToolboxApp/MainWindow.cpp

commit 9078d30d59772b128b42242df1e8744fa5037dbb
Author: SGrottel <webmaster@sgrottel.de>
Date:   Sat Jul 2 16:34:51 2022 +0200

    SettingsApp Tools now show the correct icons

commit 49620f960fc50f99ca12c86df71fabda36ded79e
Author: SGrottel <webmaster@sgrottel.de>
Date:   Sat Jul 2 16:34:29 2022 +0200

    Introduced `IconLoader` in ToolboxLib

commit f85100aec4db5e73b2bcca9d6988f6a0b4a7fdbc
Author: SGrottel <webmaster@sgrottel.de>
Date:   Sat Jul 2 14:30:30 2022 +0200

    Tool reordering and placement icon added

commit a2552ec08a70c81032b5cd7e4f84bee393597766
Author: SGrottel <webmaster@sgrottel.de>
Date:   Fri Jul 1 16:55:37 2022 +0200

    Model View ViewModel for ToolInfo

commit f6d21e5c74ddf0bcefa6ed3d9b745b45e61ebf15
Author: SGrottel <webmaster@sgrottel.de>
Date:   Fri Jul 1 10:01:41 2022 +0200

    Fundamental tool list interaction

commit a91f4bc99cf2d13e7e4969833dee8ee5678d860a
Author: SGrottel <webmaster@sgrottel.de>
Date:   Fri Jul 1 08:31:20 2022 +0200

    SettingsApp fundamental gui layout, generic settings, and most help

commit 1ef5db1b14619aa48505b06782359e6263b57a4c
Author: SGrottel <webmaster@sgrottel.de>
Date:   Fri Jun 17 15:13:33 2022 +0200

    Implemented SettingsBaseLib

commit cce374cd7aad8fd3c75c88ef57a336673a0e7daf
Author: SGrottel <webmaster@sgrottel.de>
Date:   Fri Jun 17 13:18:56 2022 +0200

    `ToolboxLib::VarStringSplitter` is now an internal class

commit 02b7dcc070907a10fc3d0d76a6b4c13bfcfefe5e
Author: SGrottel <webmaster@sgrottel.de>
Date:   Fri Jun 17 13:00:31 2022 +0200

    Version bumped to v0.4

commit 0e14dd1cfdfa71e263ec3108efa43ef9446abb00
Author: SGrottel <webmaster@sgrottel.de>
Date:   Fri Jun 17 11:57:41 2022 +0200

    Setup of SettingsApp (CSharp) and SettingsBaseLib (C++/CLI)

+++ ToolboxLib/ToolRunner.cpp
@@ -15,7 +15,7 @@
 //
 #include "Toolbox/ToolRunner.h"
 
-#include "Toolbox/VarStringSplitter.h"
+#include "VarStringSplitter.h"
 #include "Utf8Utility.h"
 
 #define WIN32_LEAN_AND_MEAN
@@ -165,7 +165,7 @@ void ToolRunner::ApplyVariables(ToolInfo::StartConfig& inOutToolStartConfig, std
 			}
 			else
 			{
-				for (int idx = 0; idx < files.size(); ++idx)
+				for (int idx = 0; idx < static_cast<int>(files.size()); ++idx)
 				{
 					std::wstring tn = L"file" + std::to_wstring(idx);
 					if (n == tn)
@@ -232,7 +232,7 @@ void ToolRunner::ApplyVariables(ToolInfo::StartConfig& inOutToolStartConfig, std
 			}
 			else
 			{
-				for (int idx = 0; idx < files.size(); ++idx)
+				for (int idx = 0; idx < static_cast<int>(files.size()); ++idx)
 				{
 					std::wstring tn = L"file" + std::to_wstring(idx);
 					if (n == tn)
@@ -279,7 +279,7 @@ void ToolRunner::ApplyVariables(ToolInfo::StartConfig& inOutToolStartConfig, std
 			}
 			else
 			{
-				for (int idx = 0; idx < files.size(); ++idx)
+				for (int idx = 0; idx < static_cast<int>(files.size()); ++idx)
 				{
 					std::wstring tn = L"filesafter" + std::to_wstring(idx);
 					if (splits[arrayArgIdx] == tn)
